Refactor client draft file structure
