+++ client/src/App.tsx
@@ -18,8 +18,8 @@ import StatsMain from "./components/StatsPage/StatsMain";
 // import StatsPlayer from "./components/StatsPage/StatsPlayer";
 import StatsSeason from "./components/StatsPage/StatsSeason";
 // import StatsTeamUI from "./components/StatsPage/StatsTeamUI";
-import CreateDraft from "./components/DraftTool/CreateDraft";
-import DraftPage from "./components/DraftTool/DraftPage";
+import CreateDraft from "./components/DraftTool/draftCreation/CreateDraft";
+import DraftPage from "./components/DraftTool/mainPages/DraftPage";
 import FearlessMain from "./components/DraftTool/mainPages/FearlessMain";
 import { FearlessProvider } from "./components/DraftTool/providers/FearlessProvider";
 import FearlessDraftPage from "./components/DraftTool/mainPages/FearlessDraftPage";
