+++ client/src/components/DraftTool/StreamView/StreamPickImage.tsx
@@ -1,6 +1,6 @@
 import { memo, useEffect, useState } from "react";
 import tempImage from "../../../assets/lblcsLogo.svg";
-import { Champion, DraftProps } from "../draftInterfaces";
+import { Champion, DraftProps } from "../interfaces/draftInterfaces";
 import { useDraftContext } from "../providers/DraftProvider";
 
 const StreamPickImage = ({
