+++ client/src/components/DraftTool/defaultStates/defaultDraftState.ts
@@ -1,4 +1,4 @@
-import { DraftProps } from "./draftInterfaces";
+import { DraftProps } from "./interfaces/draftInterfaces";
 
 export const defaultDraftState: DraftProps = {
     draftStarted: false,
