+++ client/src/components/DraftTool/draftCreation/createDraft.ts
