+++ client/src/components/DraftTool/draftRunHandlers/clientDraftHandler.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import { Socket } from "socket.io-client";
-import { DraftProps } from "./draftInterfaces";
+import { DraftProps } from "../interfaces/draftInterfaces";
 
 export const handleBanPhase = (
   socket: Socket,
