+++ client/src/components/DraftTool/draftViews/RoleSelect.tsx
@@ -1,8 +1,8 @@
-import topIcon from "../../assets/laneIcons/topIcon.svg";
-import jungleIcon from "../../assets/laneIcons/jungleIcon.svg";
-import middleIcon from "../../assets/laneIcons/middleIcon.svg";
-import bottomIcon from "../../assets/laneIcons/bottomIcon.svg";
-import supportIcon from "../../assets/laneIcons/supportIcon.svg";
+import topIcon from "../../../assets/laneIcons/topIcon.svg";
+import jungleIcon from "../../../assets/laneIcons/jungleIcon.svg";
+import middleIcon from "../../../assets/laneIcons/middleIcon.svg";
+import bottomIcon from "../../../assets/laneIcons/bottomIcon.svg";
+import supportIcon from "../../../assets/laneIcons/supportIcon.svg";
 import React, { JSX, memo } from "react";
 
 function RoleSelect({
