+++ client/src/components/DraftTool/mainPages/FearlessDraftPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { Link, useLocation, useParams } from "react-router-dom";
-import { Champion } from "../draftInterfaces";
-import DraftDisplay from "../DraftDisplay";
+import { Champion } from "../interfaces/draftInterfaces";
+import DraftDisplay from "../draftViews/DraftDisplay";
 import Button from "../../Button";
 import StreamDisplay from "../StreamView/StreamDisplay";
 import { useDraftContext } from "../providers/DraftProvider";
